Refactor open_datatree back together.
puts common parts in common.
