+++ xarray/backends/netCDF4_.py
@@ -16,6 +16,7 @@ from xarray.backends.common import (
     BackendEntrypoint,
     WritableCFDataStore,
     _normalize_path,
+    _open_datatree_netcdf,
     find_root_and_group,
     robust_getitem,
 )
@@ -675,36 +676,7 @@ class NetCDF4BackendEntrypoint(BackendEntrypoint):
     ) -> DataTree:
         from netCDF4 import Dataset as ncDataset
 
-        from xarray.backends.api import open_dataset
-        from xarray.datatree_.datatree import DataTree
-        from xarray.datatree_.datatree.treenode import NodePath
-
-        ds = open_dataset(filename_or_obj, **kwargs)
-        tree_root = DataTree.from_dict({"/": ds})
-        with ncDataset(filename_or_obj, mode="r") as ncds:
-            for path in _iter_nc_groups(ncds):
-                subgroup_ds = open_dataset(filename_or_obj, group=path, **kwargs)
-
-                # TODO refactor to use __setitem__ once creation of new nodes by assigning Dataset works again
-                node_name = NodePath(path).name
-                new_node: DataTree = DataTree(name=node_name, data=subgroup_ds)
-                tree_root._set_item(
-                    path,
-                    new_node,
-                    allow_overwrite=False,
-                    new_nodes_along_path=True,
-                )
-        return tree_root
-
-
-def _iter_nc_groups(root, parent="/"):
-    from xarray.datatree_.datatree.treenode import NodePath
-
-    parent = NodePath(parent)
-    for path, group in root.groups.items():
-        gpath = parent / path
-        yield str(gpath)
-        yield from _iter_nc_groups(group, parent=gpath)
+        return _open_datatree_netcdf(ncDataset, filename_or_obj, **kwargs)
 
 
 BACKEND_ENTRYPOINTS["netcdf4"] = ("netCDF4", NetCDF4BackendEntrypoint)
